put the form in a component and the action into another file
