+++ client/src/main.jsx
@@ -3,8 +3,8 @@ import ReactDOM from "react-dom/client";
 
 import {
   createBrowserRouter,
-  redirect,
   RouterProvider,
+  redirect,
 } from "react-router-dom";
 
 import App from "./App";
@@ -13,6 +13,8 @@ import LoginPage from "./pages/LoginPage";
 import ResultPage from "./pages/ResultPage";
 import SignUpPage from "./pages/SignUpPage";
 import PostNewOffer from "./pages/PostNewOffer";
+
+import PostOfferAction from "./actions/PostOfferAction";
 
 const ApiUrl = import.meta.env.VITE_API_URL;
 
@@ -86,42 +88,7 @@ const router = createBrowserRouter([
         id: "post-offer",
         element: <PostNewOffer />,
         loader: async () => fetch(`${ApiUrl}/api/companies`),
-        action: async ({ request }) => {
-          try {
-            const formData = await request.formData();
-
-            const jobTitle = formData.get("job_title");
-            const jobType = formData.get("job_type");
-            const content = formData.get("content");
-            const localisation = formData.get("localisation");
-            const minSalary = formData.get("min_salary");
-            const maxSalary = formData.get("max_salary");
-            const companyId = formData.get("company.id");
-
-            const response = await fetch(`${ApiUrl}/api/offers`, {
-              method: "POST",
-              headers: {
-                "Content-Type": "application/json",
-              },
-              body: JSON.stringify({
-                jobTitle,
-                jobType,
-                content,
-                localisation,
-                minSalary,
-                maxSalary,
-                companyId,
-              }),
-            });
-            if (response.ok === false) {
-              throw new Error("");
-            }
-          } catch (err) {
-            console.error("Fetch error:", err);
-            return null;
-          }
-          return redirect(`/result-page`);
-        },
+        action: PostOfferAction,
       },
       {
         path: "/login-page",
